# public/swagger.json
{
	"components": {
...
		"schemas": {
			"RegistrarEmpresaResponse": {
				"properties": {
					"msg": {
						"type": "string"
					},
					"ok": {
						"type": "boolean"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"error": {}
				},
				"required": [
					"msg",
					"ok",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FieldErrors": {
				"properties": {},
				"type": "object",
				"additionalProperties": {
					"properties": {
						"value": {},
						"message": {
							"type": "string"
						}
					},
					"required": [
						"message"
					],
					"type": "object"
				}
			},
			"ValidateError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"fields": {
						"$ref": "#/components/schemas/FieldErrors"
					}
				},
				"required": [
					"name",
					"message",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"InternalServerError": {
				"properties": {
					"msg": {
						"type": "string"
					},
					"ok": {
						"type": "boolean"
					},
					"error": {},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"msg",
					"ok",
					"error",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Persona": {
				"properties": {
					"persona_id": {
						"type": "number",
						"format": "double"
					},
					"nombre": {
						"type": "string"
					},
					"segundo_nombre": {
						"type": "string"
					},
					"primer_apellido": {
						"type": "string"
					},
					"segundo_apellido": {
						"type": "string"
					},
					"fecha_nacimiento": {
						"type": "string"
					},
					"sexo": {
						"type": "string"
					},
					"estado": {
						"type": "string"
					},
					"direccion_id": {
						"type": "number",
						"format": "double"
					},
					"cedula": {
						"type": "string"
					}
				},
				"required": [
					"nombre",
					"primer_apellido",
					"fecha_nacimiento",
					"sexo",
					"cedula"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Direccion": {
				"properties": {
					"direccion_id": {
						"type": "number",
						"format": "double"
					},
					"provincia_id": {
						"type": "number",
						"format": "double"
					},
					"municipio_id": {
						"type": "number",
						"format": "double"
					},
					"direccion": {
						"type": "string"
					},
					"codigo_postal": {
						"type": "string"
					},
					"referencia": {
						"type": "string"
					}
				},
				"required": [
					"provincia_id",
					"municipio_id",
					"direccion",
					"referencia"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Contactos": {
				"properties": {
					"contacto_id": {
						"type": "number",
						"format": "double"
					},
					"telefono": {
						"type": "string"
					},
					"movil": {
						"type": "string"
					},
					"telefono_oficina": {
						"type": "string"
					},
					"correo_electronico": {
						"type": "string"
					},
					"persona_id": {
						"type": "number",
						"format": "double"
					},
					"emp_id": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"PersonaEmpresa": {
				"properties": {
					"persona": {
						"$ref": "#/components/schemas/Persona"
					},
					"direccion": {
						"$ref": "#/components/schemas/Direccion"
					},
					"contactos": {
						"items": {
							"$ref": "#/components/schemas/Contactos"
						},
						"type": "array"
					}
				},
				"required": [
					"persona",
					"direccion",
					"contactos"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegistrarEmpresa": {
				"properties": {
					"emp_id": {
						"type": "number",
						"format": "double"
					},
					"nombre_completo": {
						"type": "string"
					},
					"nombre_corto": {
						"type": "string"
					},
					"rnc": {
						"type": "string"
					},
					"website": {
						"type": "string"
					},
					"persona": {
						"$ref": "#/components/schemas/PersonaEmpresa"
					},
					"direccion": {
						"$ref": "#/components/schemas/Direccion"
					},
					"contactos": {
						"items": {
							"$ref": "#/components/schemas/Contactos"
						},
						"type": "array"
					}
				},
				"required": [
					"nombre_completo",
					"rnc",
					"persona",
					"direccion",
					"contactos"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ConsultaEmpresasResponse": {
				"properties": {
					"ok": {
						"type": "boolean"
					},
					"data": {},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"ok",
					"data",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NotFoundItems": {
				"properties": {
					"msg": {
						"type": "string"
					},
					"ok": {
						"type": "boolean"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"msg",
					"ok",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ActualizarEmpresaResponse": {
				"properties": {
					"msg": {
						"type": "string"
					},
					"ok": {
						"type": "boolean"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"error": {}
				},
				"required": [
					"msg",
					"ok",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ActualizarEmpresa": {
				"properties": {
					"emp_id": {
						"type": "number",
						"format": "double"
					},
					"nombre_completo": {
						"type": "string"
					},
					"nombre_corto": {
						"type": "string"
					},
					"rnc": {
						"type": "string"
					},
					"website": {
						"type": "string"
					},
					"persona": {
						"$ref": "#/components/schemas/PersonaEmpresa"
					},
					"direccion": {
						"$ref": "#/components/schemas/Direccion"
					},
					"contactos": {
						"items": {
							"$ref": "#/components/schemas/Contactos"
						},
						"type": "array"
					}
				},
				"required": [
					"nombre_completo",
					"rnc",
					"persona",
					"direccion",
					"contactos"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegistrarPersonaResponse": {
				"properties": {
					"msg": {
						"type": "string"
					},
					"ok": {
						"type": "boolean"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"error": {}
				},
				"required": [
					"msg",
					"ok",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"RegistrarPersona": {
				"properties": {
					"persona": {
						"$ref": "#/components/schemas/Persona"
					},
					"contactos": {
						"items": {
							"$ref": "#/components/schemas/Contactos"
						},
						"type": "array"
					},
					"direccion": {
						"$ref": "#/components/schemas/Direccion"
					}
				},
				"required": [
					"persona",
					"contactos",
					"direccion"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ActualizarPersonaResponse": {
				"properties": {
					"msg": {
						"type": "string"
					},
					"ok": {
						"type": "boolean"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"msg",
					"ok",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PersonaInterface": {
				"properties": {
					"persona_id": {
						"type": "number",
						"format": "double"
					},
					"nombre": {
						"type": "string"
					},
					"segundo_nombre": {
						"type": "string"
					},
					"primer_apellido": {
						"type": "string"
					},
					"segundo_apellido": {
						"type": "string"
					},
					"fecha_nacimiento": {
						"type": "string"
					},
					"sexo": {
						"type": "string"
					},
					"estado": {
						"type": "string"
					},
					"direccion_id": {
						"type": "number",
						"format": "double"
					},
					"cedula": {
						"type": "string"
					}
				},
				"required": [
					"nombre",
					"primer_apellido",
					"fecha_nacimiento",
					"sexo",
					"cedula"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ActualizarPersonaBody": {
				"properties": {
					"personaId": {
						"type": "number",
						"format": "double"
					},
					"persona": {
						"$ref": "#/components/schemas/PersonaInterface"
					},
					"direccion": {
						"$ref": "#/components/schemas/Direccion"
					},
					"contactos": {
						"items": {
							"$ref": "#/components/schemas/Contactos"
						},
						"type": "array"
					}
				},
				"required": [
					"personaId",
					"persona",
					"direccion",
					"contactos"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ConsultarPersonasResponse": {
				"properties": {
					"ok": {
						"type": "boolean"
					},
					"data": {},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"ok",
					"data",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"RegistrarNoticiaResponse": {
				"properties": {
					"ok": {
						"type": "boolean"
					},
					"msg": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"ok",
					"msg",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegistrarNoticiaRequest": {
				"properties": {
					"empresa_id": {
						"type": "number",
						"format": "double"
					},
					"titulo": {
						"type": "string"
					},
					"descripcion": {
						"type": "string"
					},
					"persona_id": {
						"type": "number",
						"format": "double"
					},
					"fecha_publicacion": {
						"type": "string",
						"format": "date-time"
					},
					"fecha_vencimiento": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"empresa_id",
					"titulo",
					"descripcion",
					"persona_id",
					"fecha_publicacion",
					"fecha_vencimiento"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Noticia": {
				"properties": {
					"noticia_id": {
						"type": "number",
						"format": "double"
					},
					"empresa_id": {
						"type": "number",
						"format": "double"
					},
					"titulo": {
						"type": "string"
					},
					"descripcion": {
						"type": "string"
					},
					"persona_id": {
						"type": "number",
						"format": "double"
					},
					"fecha_publicacion": {
						"type": "string",
						"format": "date-time"
					},
					"fecha_vencimiento": {
						"type": "string",
						"format": "date-time"
					},
					"fecha_creacion": {
						"type": "string",
						"format": "date-time"
					},
					"fecha_actualizacion": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"empresa_id",
					"titulo",
					"descripcion",
					"persona_id",
					"fecha_publicacion",
					"fecha_vencimiento"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetAllNoticiasResponse": {
				"properties": {
					"ok": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/Noticia"
						},
						"type": "array"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"ok",
					"data",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateNoticiaResponse": {
				"properties": {
					"ok": {
						"type": "boolean"
					},
					"msg": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"ok",
					"msg",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateNoticiaData": {
				"properties": {
					"titulo": {
						"type": "string"
					},
					"descripcion": {
						"type": "string"
					},
					"fecha_vencimiento": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"DeleteNoticiaResponse": {
				"properties": {
					"ok": {
						"type": "boolean"
					},
					"msg": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"ok",
					"msg",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "finantek_api",
		"version": "0.0.1",
		"description": "API fir finantek",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Daniel Valdez | hello@daniel-valdez.com | https://daniel-valdez.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/empresas/registrar": {
			"post": {
				"operationId": "RegistrarEmpresas",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/RegistrarEmpresaResponse"
										},
										{
											"$ref": "#/components/schemas/ValidateError"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Empresa Existe",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "La empresa que desea registrar ya se encuentra en el sistema",
											"status": 400,
											"name": "ErrorExistEmp"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "Error interno del sistema, por favor contacte al administrador del sistema",
											"error": {},
											"status": 500
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Empresas"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegistrarEmpresa"
							}
						}
					}
				}
			}
		},
		"/api/empresas/all": {
			"get": {
				"operationId": "ConsultarEmpresas",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ConsultaEmpresasResponse"
										},
										{
											"$ref": "#/components/schemas/ValidateError"
										},
										{
											"$ref": "#/components/schemas/InternalServerError"
										},
										{
											"$ref": "#/components/schemas/NotFoundItems"
										}
									]
								}
							}
						}
					},
					"404": {
						"description": "No se encontraron empresas",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundItems"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "No se encontraron empresas con los parametros compartidos",
											"status": 404
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "Error interno del sistema, por favor contacte al administrador del sistema",
											"error": {},
											"status": 500
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Empresas"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "nombre",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "rnc",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/empresas/actualizar": {
			"put": {
				"operationId": "ActualizarEmpresa",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ActualizarEmpresaResponse"
										},
										{
											"$ref": "#/components/schemas/ValidateError"
										}
									]
								}
							}
						}
					},
					"404": {
						"description": "No se encontro la empresa",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundItems"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "No se encontro la empresa con los parametros compartidos",
											"status": 404
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "Error interno del sistema, por favor contacte al administrador del sistema",
											"error": {},
											"status": 500
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Empresas"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ActualizarEmpresa"
							}
						}
					}
				}
			}
		},
		"/api/personas/registrar": {
			"post": {
				"operationId": "RegistrarPersona",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/RegistrarPersonaResponse"
										},
										{
											"$ref": "#/components/schemas/InternalServerError"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Persona Existe",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "La persona que desea registrar ya se encuentra en el sistema",
											"status": 400,
											"name": "ErrorExistPer"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "Error interno del sistema, por favor contacte al administrador del sistema",
											"error": {},
											"status": 500
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Personas"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegistrarPersona"
							}
						}
					}
				}
			}
		},
		"/api/personas/actualizar": {
			"post": {
				"operationId": "ActualizarPersonaExistente",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ActualizarPersonaResponse"
										},
										{
											"$ref": "#/components/schemas/InternalServerError"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Persona no Existe",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "La persona que desea registrar no se encuentra en el sistema",
											"status": 400,
											"name": "ErrorExistPer"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "Error interno del sistema, por favor contacte al administrador del sistema",
											"error": {},
											"status": 500
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Personas"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ActualizarPersonaBody"
							}
						}
					}
				}
			}
		},
		"/api/personas/all": {
			"get": {
				"operationId": "GetPersonas",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ConsultarPersonasResponse"
										},
										{
											"$ref": "#/components/schemas/InternalServerError"
										},
										{
											"$ref": "#/components/schemas/NotFoundItems"
										}
									]
								}
							}
						}
					},
					"404": {
						"description": "No se encontraron personas",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundItems"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "No se encontraron personas con los parametros compartidos",
											"status": 404
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "Error interno del sistema, por favor contacte al administrador del sistema",
											"error": {},
											"status": 500
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Personas"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "nombre",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "cedula",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/personas/persona-por-id/{id}": {
			"get": {
				"operationId": "GetPersonaById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ConsultarPersonasResponse"
										},
										{
											"$ref": "#/components/schemas/NotFoundItems"
										},
										{
											"$ref": "#/components/schemas/InternalServerError"
										}
									]
								}
							}
						}
					},
					"404": {
						"description": "No se encontro la persona",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundItems"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "No se encontro la persona con los parametros compartidos",
											"status": 404
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "Error interno del sistema, por favor contacte al administrador del sistema",
											"error": {},
											"status": 500
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Personas"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/api/noticias/registrar": {
			"post": {
				"operationId": "RegistrarNoticia",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/RegistrarNoticiaResponse"
										},
										{
											"$ref": "#/components/schemas/InternalServerError"
										}
									]
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "Error interno del sistema, por favor contacte al administrador del sistema",
											"error": {},
											"status": 500
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Noticias"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegistrarNoticiaRequest"
							}
						}
					}
				}
			}
		},
		"/api/noticias/empresa/{empresa_id}": {
			"get": {
				"operationId": "GetNoticiasByEmpresa",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/GetAllNoticiasResponse"
										},
										{
											"$ref": "#/components/schemas/InternalServerError"
										}
									]
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "Error interno del sistema, por favor contacte al administrador del sistema",
											"error": {},
											"status": 500
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Noticias"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "empresa_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/noticias/update/{noticia_id}": {
			"put": {
				"operationId": "UpdateNoticia",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/UpdateNoticiaResponse"
										},
										{
											"$ref": "#/components/schemas/InternalServerError"
										}
									]
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "Error interno del sistema, por favor contacte al administrador del sistema",
											"error": {},
											"status": 500
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Noticias"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "noticia_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateNoticiaData"
							}
						}
					}
				}
			}
		},
		"/api/noticias/delete/{noticia_id}": {
			"delete": {
				"operationId": "DeleteNoticia",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/DeleteNoticiaResponse"
										},
										{
											"$ref": "#/components/schemas/InternalServerError"
										}
									]
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"ok": false,
											"msg": "Error interno del sistema, por favor contacte al administrador del sistema",
											"error": {},
											"status": 500
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Noticias"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "noticia_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}